* changes:
  feat(cbmc): add initial support for CBMC
  feat(cbmc): add host target for CBMC
  refactor(cbmc): make granule related buffers global
  feat(cbmc): add macros to conditionally compile code blocks
  fix(host): update signature of handle_ns_smc in host_rmi_wrappers.c
  refactor(cbmc): refactor platform specific constants
  refactor(lib/common): disable COMPILER_ASSERT for CBMC
  feat(cbmc): add cbmc implementations for functions in lib attestation

+++ plat/host/common/src/host_rmi_wrappers.c
@@ -6,7 +6,7 @@
 #include <host_rmi_wrappers.h>
 
 /* Declaring SMC handler */
-void handle_ns_smc(unsigned long function_id,
+void handle_ns_smc(unsigned int function_id,
 		   unsigned long arg0,
 		   unsigned long arg1,
 		   unsigned long arg2,
